+++ src/components/ViewSource/ViewSource.js
@@ -43,8 +43,8 @@ import { useGoBack } from '../../common/hooks';
 import {
   useAuditSettings,
   useQuickExport,
-  useSharedInstancesQuery,
 } from '../../hooks';
+import { useSharedInstancesQuery } from '../../Instance/hooks';
 import { IdReportGenerator } from '../../reports';
 import {
   isUserInConsortiumMode,
@@ -282,7 +282,7 @@ const ViewSource = ({
     <FormattedMessage
       id={`ui-inventory.marcSourceRecord.${marcType}`}
       values={{
-        shared: isUserInConsortiumMode(stripes) ? instance.shared : null,
+        shared: isUserInConsortiumMode(stripes) ? marcType === MARC_TYPES.BIB && instance.shared : null,
       }}
     />
   );